refactor: move Agent enum to core/models.py alongside SourceType
Central location for all enums/models. Enables future Python API use like
intpot.load('server.py').add_skills(Agent.claude) without awkward imports
from a CLI command module.
